requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
jsonschema>=4.19.0
python-dateutil>=2.8.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23
alembic>=1.13.0
PyYAML>=6.0.1
